splitting to Python and Ray
