+++ .make.versions
@@ -49,7 +49,12 @@ PROGLANG_SELECT_RAY_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
 
 CODE_QUALITY_RAY_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
 
-INPUT_CODE_TO_PARQUET_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
+INPUT_CODE_TO_PARQUET_PYTHON_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
+INPUT_CODE_TO_PARQUET_RAY_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
+
+
+
+
 INGEST_TO_PARQUET_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
