+++ data-processing-lib/python/src/data_processing/transform/__init__.py
@@ -1,3 +1,4 @@
+from data_processing.transform.abstract_transform import AbstractTransform
 from data_processing.transform.binary_transform import AbstractBinaryTransform
 from data_processing.transform.table_transform import AbstractTableTransform
 from data_processing.transform.transform_statistics import TransformStatistics
