+++ transforms/code/inputcode2parquet/python/.dockerignore
@@ -1,0 +1,1 @@
+venv/
