+++ transforms/code/inputcode2parquet/python/requirements.txt
@@ -1,0 +1,2 @@
+parameterized
+pandas
