+++ transforms/code/inputcode2parquet/python/test-data/expected/metadata.json
@@ -1,0 +1,43 @@
+{
+    "pipeline": "pipeline_id",
+    "job details": {
+        "job category": "preprocessing",
+        "job name": "ingest_to_parquet",
+        "job type": "pure python",
+        "job id": "job_id",
+        "start_time": "2024-05-31 09:00:26",
+        "end_time": "2024-05-31 09:00:55",
+        "status": "success"
+    },
+    "code": {
+        "github": "github",
+        "commit_hash": "12345",
+        "path": "path"
+    },
+    "job_input_params": {
+        "ingest_to_parquet_detect_programming_lang": true,
+        "ingest_to_parquet_supported_langs_file": "/Users/boris/Projects/data-prep-kit/transforms/code/ingest2parquet/ray/test-data/languages/lang_extensions.json",
+        "ingest_to_parquet_domain": "code",
+        "ingest_to_parquet_snapshot": "github",
+        "checkpointing": false,
+        "max_files": -1,
+        "random_samples": -1,
+        "files_to_use": [".zip"]
+    },
+    "job_output_stats": {
+        "source_files": 2,
+        "source_size": 33767275,
+        "result_files": 2,
+        "result_size": 56701,
+        "processing_time": 28.499369144439697,
+        "number of rows": 54
+    },
+    "source": {
+        "name": "/Users/boris/Projects/data-prep-kit/transforms/code/ingest2parquet/ray/test-data/input",
+        "type": "path"
+    },
+    "target": {
+        "name": "/Users/boris/Projects/data-prep-kit/transforms/code/ingest2parquet/ray/output",
+        "type": "path"
+    }
+}
