+++ transforms/code/inputcode2parquet/ray/test/test_inputcode2parquet_ray.py
@@ -13,17 +13,17 @@
 import ast
 import os
 
-from inputcode2parquet_transform_ray import (
-    CodeToParquetRayConfiguration,
+from data_processing.test_support.launch.transform_test import (
+    AbstractTransformLauncherTest,
+)
+from data_processing_ray.runtime.ray import RayTransformLauncher
+from inputcode2parquet_transform import (
     detect_programming_lang_key,
     domain_key,
     snapshot_key,
     supported_langs_file_key,
 )
-from data_processing.test_support.launch.transform_test import (
-    AbstractTransformLauncherTest,
-)
-from data_processing_ray.runtime.ray import RayTransformLauncher
+from inputcode2parquet_transform_ray import CodeToParquetRayConfiguration
 
 
 class TestRayIngestToParquetTransform(AbstractTransformLauncherTest):
