+++ transforms/universal/noop/spark/Dockerfile
@@ -14,8 +14,8 @@ RUN cd data-processing-lib-python && pip install --no-cache-dir -e .
 COPY --chown=spark:root data-processing-lib-spark/ data-processing-lib-spark/
 RUN cd data-processing-lib-spark && pip install --no-cache-dir -e .
 
-#COPY requirements.txt requirements.txt
-#RUN pip install --no-cache-dir -r requirements.txt
+COPY --chown=ray:users python-transform/  python-transform/
+RUN cd python-transform && pip install --no-cache-dir -e .
 
 COPY --chown=root:root src/ src/
 COPY --chown=root:root pyproject.toml pyproject.toml
